Refactor xtask branding command into modules
