+++ xtask/src/commands/branding/validate.rs
@@ -1,75 +1,8 @@
 use crate::error::{TaskError, TaskResult};
-use crate::util::is_help_flag;
-use crate::workspace::{WorkspaceBranding, load_workspace_branding};
-use serde_json::json;
-use std::ffi::OsString;
+use crate::workspace::WorkspaceBranding;
 use std::path::Path;
 
-/// Output format supported by the `branding` command.
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub enum BrandingOutputFormat {
-    /// Human-readable text report.
-    #[default]
-    Text,
-    /// Structured JSON report suitable for automation.
-    Json,
-}
-
-/// Options accepted by the `branding` command.
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct BrandingOptions {
-    /// Desired output format.
-    pub format: BrandingOutputFormat,
-}
-
-/// Parses CLI arguments for the `branding` command.
-pub fn parse_args<I>(args: I) -> TaskResult<BrandingOptions>
-where
-    I: IntoIterator<Item = OsString>,
-{
-    let mut options = BrandingOptions::default();
-
-    for arg in args {
-        if is_help_flag(&arg) {
-            return Err(TaskError::Help(usage()));
-        }
-
-        let Some(raw) = arg.to_str() else {
-            return Err(TaskError::Usage(String::from(
-                "branding command arguments must be valid UTF-8",
-            )));
-        };
-
-        match raw {
-            "--json" => {
-                if !matches!(options.format, BrandingOutputFormat::Text) {
-                    return Err(TaskError::Usage(String::from(
-                        "--json specified multiple times",
-                    )));
-                }
-                options.format = BrandingOutputFormat::Json;
-            }
-            _ => {
-                return Err(TaskError::Usage(format!(
-                    "unrecognised argument '{raw}' for branding command",
-                )));
-            }
-        }
-    }
-
-    Ok(options)
-}
-
-/// Executes the `branding` command.
-pub fn execute(workspace: &Path, options: BrandingOptions) -> TaskResult<()> {
-    let branding = load_workspace_branding(workspace)?;
-    validate_branding(&branding)?;
-    let output = render_branding(&branding, options.format)?;
-    println!("{output}");
-    Ok(())
-}
-
-fn validate_branding(branding: &WorkspaceBranding) -> TaskResult<()> {
+pub fn validate_branding(branding: &WorkspaceBranding) -> TaskResult<()> {
     if branding.brand.trim().is_empty() {
         return Err(TaskError::Validation(String::from(
             "workspace brand label must not be empty",
@@ -267,96 +200,6 @@ fn ensure_named_file(path: &Path, expected: &str, label: &str) -> TaskResult<()>
     Ok(())
 }
 
-fn render_branding(
-    branding: &WorkspaceBranding,
-    format: BrandingOutputFormat,
-) -> TaskResult<String> {
-    match format {
-        BrandingOutputFormat::Text => Ok(render_branding_text(branding)),
-        BrandingOutputFormat::Json => render_branding_json(branding),
-    }
-}
-
-fn render_branding_text(branding: &WorkspaceBranding) -> String {
-    format!(
-        concat!(
-            "Workspace branding summary:\n",
-            "  brand: {}\n",
-            "  upstream_version: {}\n",
-            "  rust_version: {}\n",
-            "  protocol: {}\n",
-            "  client_bin: {}\n",
-            "  daemon_bin: {}\n",
-            "  legacy_client_bin: {}\n",
-            "  legacy_daemon_bin: {}\n",
-            "  daemon_config_dir: {}\n",
-            "  daemon_config: {}\n",
-            "  daemon_secrets: {}\n",
-            "  legacy_daemon_config_dir: {}\n",
-            "  legacy_daemon_config: {}\n",
-            "  legacy_daemon_secrets: {}\n",
-            "  source: {}"
-        ),
-        branding.brand,
-        branding.upstream_version,
-        branding.rust_version,
-        branding.protocol,
-        branding.client_bin,
-        branding.daemon_bin,
-        branding.legacy_client_bin,
-        branding.legacy_daemon_bin,
-        branding.daemon_config_dir.display(),
-        branding.daemon_config.display(),
-        branding.daemon_secrets.display(),
-        branding.legacy_daemon_config_dir.display(),
-        branding.legacy_daemon_config.display(),
-        branding.legacy_daemon_secrets.display(),
-        branding.source,
-    )
-}
-
-fn render_branding_json(branding: &WorkspaceBranding) -> TaskResult<String> {
-    let value = json!({
-        "brand": branding.brand,
-        "upstream_version": branding.upstream_version,
-        "rust_version": branding.rust_version,
-        "protocol": branding.protocol,
-        "client_bin": branding.client_bin,
-        "daemon_bin": branding.daemon_bin,
-        "legacy_client_bin": branding.legacy_client_bin,
-        "legacy_daemon_bin": branding.legacy_daemon_bin,
-        "daemon_config_dir": branding.daemon_config_dir.display().to_string(),
-        "daemon_config": branding.daemon_config.display().to_string(),
-        "daemon_secrets": branding.daemon_secrets.display().to_string(),
-        "legacy_daemon_config_dir": branding
-            .legacy_daemon_config_dir
-            .display()
-            .to_string(),
-        "legacy_daemon_config": branding
-            .legacy_daemon_config
-            .display()
-            .to_string(),
-        "legacy_daemon_secrets": branding
-            .legacy_daemon_secrets
-            .display()
-            .to_string(),
-        "source": branding.source,
-    });
-
-    serde_json::to_string_pretty(&value).map_err(|error| {
-        TaskError::Metadata(format!(
-            "failed to serialise branding metadata as JSON: {error}"
-        ))
-    })
-}
-
-/// Returns usage text for the command.
-pub fn usage() -> String {
-    String::from(
-        "Usage: cargo xtask branding [--json]\n\nOptions:\n  --json          Emit branding metadata in JSON format\n  -h, --help      Show this help message",
-    )
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -384,87 +227,8 @@ mod tests {
     }
 
     fn manifest_branding() -> WorkspaceBranding {
-        let manifest = include_str!("../../../Cargo.toml");
+        let manifest = include_str!("../../../../Cargo.toml");
         parse_workspace_branding(manifest).expect("manifest parses")
-    }
-
-    #[test]
-    fn parse_args_accepts_default_configuration() {
-        let options = parse_args(std::iter::empty()).expect("parse succeeds");
-        assert_eq!(options, BrandingOptions::default());
-    }
-
-    #[test]
-    fn parse_args_reports_help_request() {
-        let error = parse_args([OsString::from("--help")]).unwrap_err();
-        assert!(matches!(error, TaskError::Help(message) if message == usage()));
-    }
-
-    #[test]
-    fn parse_args_enables_json_output() {
-        let options = parse_args([OsString::from("--json")]).expect("parse succeeds");
-        assert_eq!(
-            options,
-            BrandingOptions {
-                format: BrandingOutputFormat::Json,
-            }
-        );
-    }
-
-    #[test]
-    fn parse_args_rejects_duplicate_json_flags() {
-        let error = parse_args([OsString::from("--json"), OsString::from("--json")]).unwrap_err();
-        assert!(matches!(error, TaskError::Usage(message) if message.contains("--json")));
-    }
-
-    #[test]
-    fn parse_args_rejects_unknown_argument() {
-        let error = parse_args([OsString::from("--unknown")]).unwrap_err();
-        assert!(matches!(error, TaskError::Usage(message) if message.contains("--unknown")));
-    }
-
-    #[test]
-    fn render_text_matches_expected_layout() {
-        let branding = sample_branding();
-        let rendered = render_branding_text(&branding);
-        let expected = concat!(
-            "Workspace branding summary:\n",
-            "  brand: oc\n",
-            "  upstream_version: 3.4.1\n",
-            "  rust_version: 3.4.1-rust\n",
-            "  protocol: 32\n",
-            "  client_bin: oc-rsync\n",
-            "  daemon_bin: oc-rsyncd\n",
-            "  legacy_client_bin: rsync\n",
-            "  legacy_daemon_bin: rsyncd\n",
-            "  daemon_config_dir: /etc/oc-rsyncd\n",
-            "  daemon_config: /etc/oc-rsyncd/oc-rsyncd.conf\n",
-            "  daemon_secrets: /etc/oc-rsyncd/oc-rsyncd.secrets\n",
-            "  legacy_daemon_config_dir: /etc\n",
-            "  legacy_daemon_config: /etc/rsyncd.conf\n",
-            "  legacy_daemon_secrets: /etc/rsyncd.secrets\n",
-            "  source: https://example.invalid/rsync"
-        );
-        assert_eq!(rendered, expected);
-    }
-
-    #[test]
-    fn render_json_produces_expected_structure() {
-        let branding = sample_branding();
-        let rendered = render_branding_json(&branding).expect("json output");
-        let parsed: serde_json::Value = serde_json::from_str(&rendered).expect("parse json");
-        assert_eq!(parsed["brand"], "oc");
-        assert_eq!(parsed["protocol"], 32);
-    }
-
-    #[test]
-    fn render_branding_respects_selected_format() {
-        let branding = sample_branding();
-        let text = render_branding(&branding, BrandingOutputFormat::Text).expect("text");
-        assert_eq!(text, render_branding_text(&branding));
-        let json = render_branding(&branding, BrandingOutputFormat::Json).expect("json");
-        let expected = render_branding_json(&branding).expect("json");
-        assert_eq!(json, expected);
     }
 
     #[test]
@@ -482,7 +246,7 @@ mod tests {
     #[test]
     fn validate_branding_rejects_protocol_outside_supported_range() {
         let mut branding = sample_branding();
-        branding.protocol = 40;
+        branding.protocol = 27;
         let error = validate_branding(&branding).unwrap_err();
         assert!(matches!(
             error,
@@ -495,9 +259,10 @@ mod tests {
         let mut branding = sample_branding();
         branding.client_bin = String::from("rsync");
         let error = validate_branding(&branding).unwrap_err();
-        assert!(
-            matches!(error, TaskError::Validation(message) if message.contains("client binary"))
-        );
+        assert!(matches!(
+            error,
+            TaskError::Validation(message) if message.contains("client binary")
+        ));
     }
 
     #[test]
@@ -505,9 +270,10 @@ mod tests {
         let mut branding = sample_branding();
         branding.daemon_bin = String::from("rsyncd");
         let error = validate_branding(&branding).unwrap_err();
-        assert!(
-            matches!(error, TaskError::Validation(message) if message.contains("daemon binary"))
-        );
+        assert!(matches!(
+            error,
+            TaskError::Validation(message) if message.contains("daemon binary")
+        ));
     }
 
     #[test]
@@ -515,7 +281,10 @@ mod tests {
         let mut branding = sample_branding();
         branding.brand.clear();
         let error = validate_branding(&branding).unwrap_err();
-        assert!(matches!(error, TaskError::Validation(message) if message.contains("brand label")));
+        assert!(matches!(
+            error,
+            TaskError::Validation(message) if message.contains("must not be empty")
+        ));
     }
 
     #[test]
@@ -574,7 +343,8 @@ mod tests {
         let error = validate_branding(&branding).unwrap_err();
         assert!(matches!(
             error,
-            TaskError::Validation(message) if message.contains("legacy_daemon_config_dir")
+            TaskError::Validation(message)
+                if message.contains("legacy_daemon_config_dir")
         ));
     }
 
@@ -585,7 +355,8 @@ mod tests {
         let prefix_error = validate_branding(&branding).unwrap_err();
         assert!(matches!(
             prefix_error,
-            TaskError::Validation(message) if message.contains("must include upstream_version")
+            TaskError::Validation(message)
+                if message.contains("must include upstream_version")
         ));
 
         let mut branding = sample_branding();
